refactor: useTime훅 useActiveTime과 useActiveDate로 분리 / useActiveTime 훅 리팩토링
